frontend/requirements.txt
Flask==2.0.3
psycopg2-binary
Werkzeug==2.0.3
